- added a mersenne twister basedn prng
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid-7.2@982 82007160-df01-0410-b94d-b575c5fd34c7

+++ modules/m_pong.c
@@ -109,7 +109,7 @@ mr_pong(struct Client *client_p, struct Client *source_p,
   {
     if (ConfigFileEntry.ping_cookie && !source_p->localClient->registration)
     {
-      unsigned long incoming_ping = strtoul(parv[1], NULL, 10);
+      unsigned int incoming_ping = strtoul(parv[1], NULL, 10);
 
       if (incoming_ping)
       {
